WIP: Added a bit more logic. Made methods async.

+++ FireProhobition.Lib/Model/Municipality.cs
@@ -4,7 +4,7 @@
     {
         public required string Code { get; set; }
         public required string Name { get; set; }
-        public required float Latitude { get; set; }
-        public required float Longitude { get; set; }
+        public required double Latitude { get; set; }
+        public required double Longitude { get; set; }
     }
 }
